fix #280771 midi entry on empty voice

+++ libmscore/cmd.cpp
@@ -2001,7 +2001,7 @@ bool Score::processMidiInput()
                         ev.chord = true;
 
                   Element* cr = _is.lastSegment()->element(_is.track());
-                  if (cr->isChord()) {
+                  if (cr && cr->isChord()) {
                         Note* n = toChord(cr)->findNote(ev.pitch);
                         if (n) {
                               deleteItem(n->tieBack());
